// src/cst-definitions.d.ts
export interface ExprCstNode extends CstNode {
  name: "expr";
  children: ExprCstChildren;
}
// ...
export interface IdentifierExpressionCstNode extends CstNode {
  name: "identifierExpression";
  children: IdentifierExpressionCstChildren;
}

export type IdentifierExpressionCstChildren = {
  Identifier: IToken[];
  identifierDotExpression?: IdentifierDotExpressionCstNode[];
  identifierIndexExpression?: IndexExpressionCstNode[];
};

export interface IdentifierDotExpressionCstNode extends CstNode {
  name: "identifierDotExpression";
  children: IdentifierDotExpressionCstChildren;
}

export type IdentifierDotExpressionCstChildren = {
  Dot: IToken[];
  Identifier: IToken[];
  OpenParenthesis?: IToken[];
  arg?: ExprCstNode[];
  Comma?: IToken[];
  args?: ExprCstNode[];
  CloseParenthesis?: IToken[];
};

export interface IndexExpressionCstNode extends CstNode {
  name: "indexExpression";
  children: IndexExpressionCstChildren;
}

export type IndexExpressionCstChildren = {
  OpenBracket: IToken[];
  expr: ExprCstNode[];
  CloseBracket: IToken[];
};